// src/uikit/widgets/Menu/types.ts
import { ReactElement } from 'react'

export interface Language {
  code: string
  language: string
  locale: string
}

export interface PushedProps {
  isPushed: boolean
  pushNav: (isPushed: boolean) => void
}

export interface NavTheme {
  background: string
}

export interface LinkStatus {
  text: string
  color: string
}

export interface MenuSubEntry {
  label: string
  href: string
  calloutClass?: string
  status?: LinkStatus
}

export interface MenuEntry {
  label: string
  icon: any
  items?: MenuSubEntry[]
  href?: string
  calloutClass?: string
  initialOpenState?: boolean
  status?: LinkStatus
}

export interface PanelProps {
  isDark: boolean
  toggleTheme: (isDark: boolean) => void
  cakePriceUsd?: number
  currentLang: string
  langs: Language[]
  setLang: (lang: Language) => void
  links: Array<MenuEntry>
}

export interface NavProps extends PanelProps {
  userMenu?: ReactElement
}